+++ src/com/example/wc_tool/interfaces/WcToolOperations.java
@@ -4,9 +4,9 @@ import com.example.wc_tool.objects.FileClass;
 
 public interface WcToolOperations {
 
-    int countBytes(FileClass fc);
-    int countLines(FileClass fc);
-    int countWords(FileClass fc);
-    int countCharacters(FileClass fc);
+    long countBytes(FileClass fc);
+    long countLines(FileClass fc);
+    long countWords(FileClass fc);
+    long countCharacters(FileClass fc);
 
 }
